Kommentarer, borttagning av slumpade värden
Skrivit massor av kommentarer, samt testat ta bort slumpmässiga värden
