+++ src/priorityAlgorithm.hpp
@@ -1,0 +1,8 @@
+#ifndef INTEGER_CLASS_HPP
+#define INTEGER_CLASS_HPP
+
+// Method to calculate priority and return an integer value that represents a priority status
+int calculatePriority(float battery_charge, float min_charge);
+
+
+#endif // INTEGER_CLASS_HPP
